CloudWatch: use new QueryHeader in Logs
