--- .yarn/sdks/stylelint/lib/index.js
@@ -1,20 +1,0 @@
-#!/usr/bin/env node
-
-const {existsSync} = require(`fs`);
-const {createRequire, createRequireFromPath} = require(`module`);
-const {resolve} = require(`path`);
-
-const relPnpApiPath = "../../../../.pnp.cjs";
-
-const absPnpApiPath = resolve(__dirname, relPnpApiPath);
-const absRequire = (createRequire || createRequireFromPath)(absPnpApiPath);
-
-if (existsSync(absPnpApiPath)) {
-  if (!process.versions.pnp) {
-    // Setup the environment to be able to require stylelint/lib/index.js
-    require(absPnpApiPath).setup();
-  }
-}
-
-// Defer to the real stylelint/lib/index.js your application uses
-module.exports = absRequire(`stylelint/lib/index.js`);
