+++ public/app/plugins/datasource/cloudwatch/components/QueryHeader.tsx
@@ -1,22 +1,16 @@
-import React, { useCallback, useState } from 'react';
+import React from 'react';
 import { pick } from 'lodash';
 
-import { SelectableValue } from '@grafana/data';
-import { Button, ConfirmModal, RadioButtonGroup } from '@grafana/ui';
-import { EditorHeader, InlineSelect, FlexItem } from '@grafana/experimental';
+import { ExploreMode, SelectableValue } from '@grafana/data';
+import { EditorHeader, InlineSelect } from '@grafana/experimental';
 
 import { CloudWatchDatasource } from '../datasource';
-import {
-  CloudWatchMetricsQuery,
-  CloudWatchQuery,
-  CloudWatchQueryMode,
-  MetricEditorMode,
-  MetricQueryType,
-} from '../types';
+import { CloudWatchQuery, CloudWatchQueryMode } from '../types';
 import { useRegions } from '../hooks';
+import MetricsQueryHeader from './MetricsQueryHeader';
 
 interface QueryHeaderProps {
-  query: CloudWatchMetricsQuery;
+  query: CloudWatchQuery;
   datasource: CloudWatchDatasource;
   onChange: (query: CloudWatchQuery) => void;
   onRunQuery: () => void;
@@ -28,36 +22,10 @@ const apiModes: Array<SelectableValue<CloudWatchQueryMode>> = [
   { label: 'CloudWatch Logs', value: 'Logs' },
 ];
 
-const metricEditorModes: Array<SelectableValue<MetricQueryType>> = [
-  { label: 'Metric Search', value: MetricQueryType.Search },
-  { label: 'Metric Query', value: MetricQueryType.Query },
-];
-
-const editorModes = [
-  { label: 'Builder', value: MetricEditorMode.Builder },
-  { label: 'Code', value: MetricEditorMode.Code },
-];
-
 const QueryHeader: React.FC<QueryHeaderProps> = ({ query, sqlCodeEditorIsDirty, datasource, onChange, onRunQuery }) => {
-  const { metricEditorMode, metricQueryType, queryMode, region } = query;
-  const [showConfirm, setShowConfirm] = useState(false);
+  const { queryMode, region } = query;
 
   const [regions, regionIsLoading] = useRegions(datasource);
-
-  const onEditorModeChange = useCallback(
-    (newMetricEditorMode: MetricEditorMode) => {
-      if (
-        sqlCodeEditorIsDirty &&
-        metricQueryType === MetricQueryType.Query &&
-        metricEditorMode === MetricEditorMode.Code
-      ) {
-        setShowConfirm(true);
-        return;
-      }
-      onChange({ ...query, metricEditorMode: newMetricEditorMode });
-    },
-    [setShowConfirm, onChange, sqlCodeEditorIsDirty, query, metricEditorMode, metricQueryType]
-  );
 
   const onQueryModeChange = ({ value }: SelectableValue<CloudWatchQueryMode>) => {
     if (value !== queryMode) {
@@ -66,7 +34,7 @@ const QueryHeader: React.FC<QueryHeaderProps> = ({ query, sqlCodeEditorIsDirty, 
       onChange({
         ...commonProps,
         queryMode: value,
-      });
+      } as CloudWatchQuery);
     }
   };
 
@@ -84,38 +52,15 @@ const QueryHeader: React.FC<QueryHeaderProps> = ({ query, sqlCodeEditorIsDirty, 
 
       <InlineSelect aria-label="Query mode" value={queryMode} options={apiModes} onChange={onQueryModeChange} />
 
-      <InlineSelect
-        aria-label="Metric editor mode"
-        value={metricEditorModes.find((m) => m.value === metricQueryType)}
-        options={metricEditorModes}
-        onChange={({ value }) => {
-          onChange({ ...query, metricQueryType: value });
-        }}
-      />
-
-      <FlexItem grow={1} />
-
-      <RadioButtonGroup options={editorModes} size="sm" value={metricEditorMode} onChange={onEditorModeChange} />
-
-      {query.metricQueryType === MetricQueryType.Query && query.metricEditorMode === MetricEditorMode.Code && (
-        <Button variant="secondary" size="sm" onClick={() => onRunQuery()}>
-          Run query
-        </Button>
+      {queryMode !== ExploreMode.Logs && (
+        <MetricsQueryHeader
+          query={query}
+          datasource={datasource}
+          onChange={onChange}
+          onRunQuery={onRunQuery}
+          sqlCodeEditorIsDirty={sqlCodeEditorIsDirty}
+        />
       )}
-
-      <ConfirmModal
-        isOpen={showConfirm}
-        title="Are you sure?"
-        body="You will lose manual changes done to the query if you go back to the visual builder."
-        confirmText="Yes, I am sure."
-        dismissText="No, continue editing the query manually."
-        icon="exclamation-triangle"
-        onConfirm={() => {
-          setShowConfirm(false);
-          onChange({ ...query, metricEditorMode: MetricEditorMode.Builder });
-        }}
-        onDismiss={() => setShowConfirm(false)}
-      />
     </EditorHeader>
   );
 };
